requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file requirements/base.txt requirements/base.in
#
certifi==2018.10.15       # via requests
chardet==3.0.4            # via requests
django-appconf==1.0.2     # via django-axes
django-axes==4.4.2
django-choices==1.6.1
django-ipware==2.1.0      # via django-axes
django-sniplates==0.6.0
django-solo==1.1.3
django==2.1.3
gemma-zds-client==0.6.1
idna==2.7                 # via requests
psycopg2==2.7.6.1
pyjwt==1.6.4
pytz==2018.7
pyyaml==3.13              # via gemma-zds-client
raven==6.9.0
requests==2.20.1
urllib3==1.24.1           # via requests
